+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.9] - 2021-08-10
+
+### Added
+
+- Add open graph customizer inputs for image and description for post archive
+
+### Changed
+
+- Changed title in index.php to the_archive_title for filtering
+- Replace "allowed_block_types" filter with "allowed_block_types_all"
+
+### Fixed
+
+- Fixed Fontsize and Lineheight within the block editor
+- Fixed PHP Notice on new Widget Block Editor Page
+
 ## [1.0.8] - 2021-07-22
 
 ### Added